% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-auth.R
\name{bigrquery-deprecated}
\alias{bigrquery-deprecated}
\alias{get_access_cred}
\alias{set_access_cred}
\alias{reset_access_cred}
\alias{has_access_cred}
\alias{set_oauth2.0_cred}
\alias{set_service_token}
\title{Deprecated functions for access credentials}
\usage{
get_access_cred()

set_access_cred(value)

reset_access_cred()

has_access_cred()

set_oauth2.0_cred(app)

set_service_token(service_token)
}
\arguments{
\item{value}{New access credentials, as returned by \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}.}

\item{app}{A Google OAuth application created with \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{service_token}{A JSON string, URL or file, providing a service account
token.}
}
\value{
\code{get_access_cred()} returns the current OAuth2 credentials token.
}
\description{
bigrquery's main functions for managing credentials are now \code{\link[=bq_auth]{bq_auth()}} and
\code{\link[=bq_auth_configure]{bq_auth_configure()}}. The functions documented here are now deprecated and
will eventually be removed.
}
\keyword{internal}
