% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrquery-package.R
\docType{package}
\name{bigrquery-package}
\alias{bigrquery}
\alias{bigrquery-package}
\title{bigrquery: An Interface to Google's 'BigQuery' 'API'}
\description{
Easily talk to Google's 'BigQuery' database from R.
}
\section{Package options}{

\describe{
\item{\code{bigrquery.quiet}}{Verbose output during processing? The default
value, \code{NA}, turns on verbose output for queries that run longer than
two seconds.  Use \code{FALSE} for immediate verbose output, \code{TRUE}
for quiet operation.}

\item{\code{bigrquery.page.size}}{Default page size for fetching data,
defaults to 1e4.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://bigrquery.r-dbi.org}
  \item \url{https://github.com/r-dbi/bigrquery}
  \item Report bugs at \url{https://github.com/r-dbi/bigrquery/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
  \item Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
