% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrquery.r
\docType{package}
\name{bigrquery}
\alias{bigquery}
\alias{bigrquery}
\alias{bigrquery-package}
\title{An R interface to google's bigquery.}
\description{
For more information about how bigrquery works, and how to get started,
please see the project development page: \url{github.com/rstats-db/bigrquery}.
The most important method to get started with is \code{\link{query_exec}}.
}

