% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-param.R
\name{bq-param}
\alias{bq-param}
\alias{bq_param_scalar}
\alias{bq_param_array}
\title{Defines parameter that can be passed to the query}
\usage{
bq_param_scalar(value, name = NULL, type = NULL)

bq_param_array(value, name = NULL, type = NULL)
}
\arguments{
\item{value}{vector of parameter values}

\item{name}{name of the parameter in the query, omitting the \code{@}}

\item{type}{BigQuery type of the parameter}
}
\description{
Defines parameter that can be passed to the query
}
