% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-dataset.R
\name{bq_dataset}
\alias{bq_dataset}
\alias{as_bq_dataset}
\title{BigQuery dataset class}
\usage{
bq_dataset(project, dataset)

as_bq_dataset(x)
}
\arguments{
\item{project, dataset}{Project and dataset identifiers.}

\item{x}{An object to coerce to a \code{bq_dataset}}
}
\description{
\code{bq_dataset()} will create a reference to dataset from the \code{project} and
\code{dataset} identifiers; \code{as_bq_dataset()} will create a dataset from a string.
See \link{dataset-API} for the things you can do with a dataset.
}
\examples{
bq_dataset("publicdata", "shakespeare")
as_bq_dataset("publicdata.shakespeare")
}
