% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_endpoints.R
\name{bq_endpoints}
\alias{bq_endpoints}
\title{List BigQuery endpoints}
\usage{
bq_endpoints(i = NULL)
}
\arguments{
\item{i}{The name(s) or integer index(ices) of the endpoints to return.
Optional. By default, the entire list is returned.}
}
\value{
A list containing some or all of the BigQuery API v2 endpoints that
are used internally by bigrquery.
}
\description{
Returns a list of BigQuery API v2 endpoints, as stored inside the bigrquery
package. The names of this list (or the \code{id} sub-elements) are the nicknames
that can be used to specify an endpoint in \code{\link[=request_generate]{request_generate()}}. For each
endpoint, we store its nickname or \code{id}, the associated HTTP verb, the
\code{path}, and details about the parameters. This list is derived
programmatically from the \href{https://www.googleapis.com/discovery/v1/apis/bigquery/v2/rest}{BigQuery API v2 Discovery Document}.
Vocabulary note: we use the term "endpoints", but Google's API docs usually
refer to these as "methods".
}
\examples{
names(bq_endpoints())
str(bq_endpoints(), max.level = 2)
bq_endpoints("bigquery.tables.get")
}
