% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-job.r
\name{bq_job}
\alias{bq_job}
\alias{as_bq_job}
\title{BigQuery job class}
\usage{
bq_job(project, job)

as_bq_job(x)
}
\arguments{
\item{project, job}{Project and job identifiers}

\item{x}{An object to coerce to a job; usually a string.}
}
\description{
BigQuery job class
}
\examples{
\dontrun{
bq_job("bigrquery-examples", "m0SgFu2ycbbge6jgcvzvflBJ_Wft")
as_bq_job("bigrquery-examples:US.job_m0SgFu2ycbbge6jgcvzvflBJ_Wft")
}
}
