% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{bq_project_datasets}
\alias{bq_project_datasets}
\alias{list_datasets}
\title{List the datasets in a project}
\usage{
bq_project_datasets(project, page_size = 50, max_pages = Inf)
}
\arguments{
\item{project}{A project identifier}

\item{page_size}{Number of items per page}

\item{max_pages}{Maximum number of pages to retrieve}
}
\value{
a character vector of dataset names
}
\description{
List the datasets in a project
}
\examples{
if (bq_authable()) {
bq_project_datasets("publicdata")
bq_project_datasets("githubarchive")
}
}
\seealso{
\href{https://developers.google.com/bigquery/docs/reference/v2/datasets/list}{API documentation}
}
