% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{bq_table}
\alias{bq_table}
\alias{as_bq_table}
\title{Create a table reference}
\usage{
bq_table(project, dataset, table)

as_bq_table(x)
}
\description{
\code{bq_table()} creates from the individual components; \code{as_bq_table()}
will create from a character vector or a list.
}
\examples{
natality <- bq_table("publicdata", "samples", "natality")
natality
get_table(natality)

as_bq_table("publicdata:samples:natality")
as_bq_table("publicdata.samples.natality")

# Included primarily for backward compatbility
as_bq_table(list(
  project_id = "publicdata",
  dataset_id = "sameples",
  table_id = "natality"
))
}
