% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-test.R
\name{bq_test_project}
\alias{bq_test_project}
\alias{bq_test_dataset}
\alias{bq_testable}
\alias{bq_authable}
\title{Project to use for testing bigquery}
\usage{
bq_test_project()

bq_test_dataset()

bq_testable()

bq_authable()
}
\value{
\code{bq_test_project()} returns the name of a project suitable for
use in testing. \code{bq_test_dataset()} creates a temporary dataset
who's lifetime is tied to the lifetime of the object that it returns.
}
\description{
You'll need to set up the \code{BIGQUERY_TEST_PROJECT} env var if you want to
run bigquery tests locally. I recommend creating a new project because
the tests involve both reading and writing. You will also need to
have billing billing enabled for the project, and to run \code{bq_test_init()}
once.
}
\section{Testing}{

In tests, \code{bq_test_project()} (and hence \code{bq_test_dataset()}) will
automatically skip if auth and a test project are not available.
}

\examples{
if (bq_testable()) {
  ds <- bq_test_dataset()
  bq_mtcars <- bq_table_upload(bq_table(ds, "mtcars"), mtcars)

  # dataset and table will be automatically deleted when ds is GC'd
}
}
\keyword{internal}
