% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-tranfer-config.R
\name{bq_transfer_config_create}
\alias{bq_transfer_config_create}
\title{Manage TransferConfigs}
\usage{
bq_transfer_config_create(
  project_id,
  name,
  data_source_id,
  destination_dataset_id,
  location_id = NULL,
  schedule = "",
  display_name = NULL,
  params = list(),
  schedule_options = list(),
  data_refresh_window_days = NULL,
  disabled = FALSE,
  ...
)
}
\arguments{
\item{data_source_id}{Data source ID. This cannot be changed once data transfer is created. The full list of available data source IDs can be returned through an API call: \url{https://cloud.google.com/bigquery-transfer/docs/reference/datatransfer/rest/v1/projects.locations.dataSources/list}}

\item{destination_dataset_id}{The BigQuery target dataset id.}

\item{schedule}{Data transfer schedule. If the data source does not support a custom schedule, this should be empty. If it is empty, the default value for the data source will be used. The specified times are in UTC. Examples of valid format: \verb{1st,3rd monday of month 15:30}, \verb{every wed,fri of jan,jun 13:15}, and \verb{first sunday of quarter 00:00}. See more explanation about the format here: \url{https://cloud.google.com/appengine/docs/flexible/python/scheduling-jobs-with-cron-yaml#the_schedule_format} NOTE: The minimum interval time between recurring transfers depends on the data source; refer to the documentation for your data source.}

\item{display_name}{User specified display name for the data transfer.}

\item{params}{Parameters specific to each data source. For more information see the bq tab in the 'Setting up a data transfer' section for each data source. For example the parameters for Cloud Storage transfers are listed here: \url{https://cloud.google.com/bigquery-transfer/docs/cloud-storage-transfer#bq}}

\item{schedule_options}{Options customizing the data transfer schedule.}

\item{data_refresh_window_days}{The number of days to look back to automatically refresh the data. For example, if \code{data_refresh_window_days = 10}, then every day BigQuery reingests data for \link{today-10, today-1}, rather than ingesting data for just \link{today-1}. Only valid if the data source supports the feature. Set the value to 0 to use the default value.}

\item{disabled}{Is this config disabled. When set to true, no runs are scheduled for a given transfer.}

\item{...}{Additional arguments are automatically converted to camelCamel
and added to the TransferConfig object.}
}
\description{
Manage TransferConfigs
}
