% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bqs_ua}
\alias{bqs_ua}
\title{Loosely adapted from https://github.com/googleapis/python-api-core/blob/master/google/api_core/client_info.py}
\usage{
bqs_ua()
}
\description{
Loosely adapted from https://github.com/googleapis/python-api-core/blob/master/google/api_core/client_info.py
}
