% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{copy_to.src_bigquery}
\alias{copy_to.src_bigquery}
\title{Copy a local data frame to a bigquery src through \code{\link{insert_upload_job}}.}
\usage{
\method{copy_to}{src_bigquery}(dest, df, name = deparse(substitute(df)),
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_APPEND")
}
\arguments{
\item{dest}{remote data source}

\item{df}{local data frame}

\item{name}{name for new remote table.}

\item{create_disposition}{behavior for table creation if the destination
already exists. defaults to \code{"CREATE_IF_NEEDED"},
the only other supported value is \code{"CREATE_NEVER"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.createDisposition}{the API documentation}
for more information}

\item{write_disposition}{behavior for writing data if the destination already
exists. defaults to \code{"WRITE_APPEND"}, other possible values are
\code{"WRITE_TRUNCATE"} and \code{"WRITE_EMPTY"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.writeDisposition}{the API documentation}
for more information}
}
\value{
a bigquery \code{\link{tbl}} object
}
\description{
Copy a local data frame to a bigquery src through \code{\link{insert_upload_job}}.
}
\seealso{
\code{\link{insert_upload_job}}
}

