% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.r
\name{count_tables}
\alias{count_tables}
\title{Count the total number of available tables in dataset.}
\usage{
count_tables(project, dataset)
}
\arguments{
\item{project}{project containing this table}

\item{dataset}{dataset containing this table}
}
\value{
the total number of tables
}
\description{
Count the total number of available tables in dataset.
}
\examples{
\dontrun{
count_tables("publicdata", "samples")
count_tables("githubarchive", "github")
}
}
\seealso{
API documentation:
  \url{https://developers.google.com/bigquery/docs/reference/v2/tables/list}

Other tables: \code{\link{delete_table}};
  \code{\link{get_table}}; \code{\link{list_tables}}
}

