% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-dataset.R
\name{dataset-API}
\alias{dataset-API}
\alias{bq_dataset_meta}
\alias{bq_dataset_exists}
\alias{bq_dataset_delete}
\alias{bq_dataset_create}
\alias{bq_dataset_udpate}
\alias{bq_dataset_tables}
\title{Manipulate BigQuery datasets}
\usage{
bq_dataset_meta(x, fields = NULL)

bq_dataset_exists(x)

bq_dataset_delete(x, recursive = FALSE)

bq_dataset_create(x, ...)

bq_dataset_udpate(x, ...)

bq_dataset_tables(x, page_size = 50, max_pages = Inf, ...)
}
\arguments{
\item{x}{A \link{bq_dataset}}

\item{recursive}{If \code{TRUE}, will recursively delete all contents of the dataset.}

\item{...}{Additional arguments merged into the body of the
request. \code{snake_case} will automatically be converted into
\code{camelCase} so you can use consistent argument names.}

\item{page_size}{Number of items per page}

\item{max_pages}{Maximum number of pages to retrieve}
}
\description{
Manipulate BigQuery datasets
}
\section{API documentation}{

\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/v2/datasets/get}{get}
\item \href{https://cloud.google.com/bigquery/docs/reference/v2/datasets/insert}{insert}
\item \href{https://cloud.google.com/bigquery/docs/reference/v2/datasets/delete}{delete}
\item \href{https://cloud.google.com/bigquery/docs/reference/v2/tables/list}{list}
}
}

\examples{
if (bq_testable()) {
ds <- bq_dataset(bq_test_project(), "dataset_api")
bq_dataset_exists(ds)

bq_dataset_create(ds)
bq_dataset_exists(ds)
str(bq_dataset_meta(ds))

bq_dataset_delete(ds)
bq_dataset_exists(ds)

# Use bq_test_dataset() to create a temporary dataset that will
# be automatically deleted
ds <- bq_test_dataset()
bq_table_create(bq_table(ds, "x1"))
bq_table_create(bq_table(ds, "x2"))
bq_table_create(bq_table(ds, "x3"))
bq_dataset_tables(ds)
}
}
