% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-dataset.R
\name{dataset-dep}
\alias{dataset-dep}
\alias{get_dataset}
\alias{exists_dataset}
\alias{delete_dataset}
\alias{insert_dataset}
\alias{update_dataset}
\alias{list_tables}
\title{Dataset API \link{deprecated}}
\usage{
get_dataset(project, dataset)

exists_dataset(project, dataset)

delete_dataset(project, dataset, deleteContents = FALSE)

insert_dataset(project, dataset, ...)

update_dataset(project, dataset, ...)

list_tables(project, dataset, page_size = 50, max_pages = Inf)
}
\description{
Please use \link{api-dataset} instead.
}
\keyword{internal}
