% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-driver.r
\name{dbi_driver}
\alias{dbi_driver}
\title{BigQuery DBI driver}
\usage{
dbi_driver()
}
\description{
Creates a BigQuery DBI driver for use in \code{\link[DBI]{dbConnect}}.
}
\examples{
\dontrun{
#' library(DBI)
dbConnect(dbi_driver(), database = "mydb", project = "myproject")
}
}

