% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{fetch_oauth2_creds}
\alias{fetch_oauth2_creds}
\title{Set the default mechanism for fetching OAuth2 tokens, namely the
"three-legged OAuth dance" (3LO). This simply asks the user to
authenticate our app via a browser.}
\usage{
fetch_oauth2_creds()
}
\description{
Set the default mechanism for fetching OAuth2 tokens, namely the
"three-legged OAuth dance" (3LO). This simply asks the user to
authenticate our app via a browser.
}

