% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{fetch_service_token_creds}
\alias{fetch_service_token_creds}
\title{Fetches a valid OAuth token using the previously set service
token file via the "two-legged OAuth dance" (2LO).}
\usage{
fetch_service_token_creds()
}
\description{
Fetches a valid OAuth token using the previously set service
token file via the "two-legged OAuth dance" (2LO).
}

