% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{format_dataset}
\alias{format_dataset}
\title{Format dataset and project ID as a BQ-style identifier}
\usage{
format_dataset(project_id, dataset)
}
\arguments{
\item{project_id}{project ID}

\item{dataset}{dataset name}
}
\value{
a character.
}
\description{
Format dataset and project ID as a BQ-style identifier
}

