% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{format_table}
\alias{format_table}
\title{Format dataset, project and table ID as a BQ-style identifier}
\usage{
format_table(project_id, dataset, table)
}
\arguments{
\item{project_id}{project ID}

\item{dataset}{dataset name}

\item{table}{table ID}
}
\value{
a character.
}
\description{
This function composes a table identifier from its individual
components.
}
\seealso{
Other identifier functions: \code{\link{format_dataset}},
  \code{\link{parse_dataset}}, \code{\link{parse_table}}
}
