\name{get_job}
\alias{get_job}
\title{Check status of a job.}
\usage{
  get_job(project, job)
}
\arguments{
  \item{project}{project name}

  \item{job}{job id}
}
\value{
  a job resource list, as documented at
  \url{https://developers.google.com/bigquery/docs/reference/v2/jobs}
}
\description{
  Check status of a job.
}
\seealso{
  API documentation for get method:
  \url{https://developers.google.com/bigquery/docs/reference/v2/jobs/get}
}

