% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_records}
\alias{get_records}
\title{Show RECORD: repeated columns in a BigQuery table}
\usage{
get_records(con, tbl_name)
}
\arguments{
\item{con}{BigQueryConnection object}

\item{tbl_name}{Character string of BigQuery dataset.table}
}
\description{
BigQuery can store data in RECORD: repeated columns in a table. These are structured arrays of data function as nested data frames. RECORD: repeated columns cannot be queried directly and must be unnested using BigQuery's \code{UNNEST()} function.
}
\examples{

get_records(con, "eg_dataset.eg_table")
}
