% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-id.R
\name{id-dep}
\alias{id-dep}
\alias{parse_dataset}
\alias{format_dataset}
\alias{parse_table}
\alias{format_table}
\title{Deprecated table/dataset objects}
\usage{
parse_dataset(dataset, project_id = NULL)

format_dataset(project_id, dataset)

parse_table(table, project_id = NULL)

format_table(project_id, dataset, table)
}
\description{
Please see \link{bq_table} and \link{bq_dataset} instead.
}
\keyword{internal}
