% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{insert_extract_job}
\alias{insert_extract_job}
\title{Create a new extract job.}
\usage{
insert_extract_job(source_table, project, destinationUris,
  compression = "NONE", destinationFormat = "CSV", fieldDelimiter = ",",
  printHeader = TRUE)
}
\arguments{
\item{source_table}{Source table,
either as a string in the format used by BigQuery, or as a list with
\code{project_id}, \code{dataset_id}, and \code{table_id} entries}

\item{project}{project name}

\item{destinationUris}{Specify the extract destination URI. Note: for large files, you may need to specify a wild-card since}

\item{compression}{Compression type ("NONE", "GZIP")}

\item{destinationFormat}{Destination format}

\item{fieldDelimiter}{Field delimiter (Default = ",")}

\item{printHeader}{Whether to print out a header row in the results. (Default is true)}
}
\value{
a job resource list, as documented at
  \url{https://cloud.google.com/bigquery/docs/reference/v2/jobs}
}
\description{
This is a low-level function that creates an extract job. To wait until it is
finished, see \code{\link{extract_exec}}
}
\seealso{
Other jobs: \code{\link{get_job}},
  \code{\link{insert_query_job}},
  \code{\link{insert_upload_job}}, \code{\link{wait_for}}
}

