% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{insert_load_from_cs_job}
\alias{insert_load_from_cs_job}
\title{Load data from Google Cloud Storage}
\usage{
insert_load_from_cs_job(project, dataset, table, billing = project,
  source_uris, create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_APPEND")
}
\arguments{
\item{project}{project containing this table}

\item{dataset}{dataset containing this table}

\item{table}{name of table to load data into}

\item{billing}{project ID to use for billing}

\item{source_uris}{the fully-qualified URIs that point to your data in Google
Cloud Storage; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.sourceUris}{the API documentation}
for more information}

\item{create_disposition}{behavior for table creation if the destination
already exists. defaults to \code{"CREATE_IF_NEEDED"},
the only other supported value is \code{"CREATE_NEVER"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.createDisposition}{the API documentation}
for more information}

\item{write_disposition}{behavior for writing data if the destination already
exists. defaults to \code{"WRITE_APPEND"}, other possible values are
\code{"WRITE_TRUNCATE"} and \code{"WRITE_EMPTY"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.writeDisposition}{the API documentation}
for more information}
}
\description{
Load data from Google Cloud Storage
}
\seealso{
Other jobs: \code{\link{get_job}},
  \code{\link{insert_query_job}},
  \code{\link{insert_upload_job}}, \code{\link{wait_for}}
}

