% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_query_job}
\alias{insert_query_job}
\title{Create a new query job.}
\usage{
insert_query_job(project, dataset, query, billing = project,
  destination = NULL)
}
\arguments{
\item{project}{project name}

\item{dataset}{dataset name}

\item{query}{SQL query string}

\item{billing}{project to bill to, if different to \code{project}}

\item{destination}{(optional) destination table for large queries}
}
\value{
a job resource list, as documented at
  \url{https://developers.google.com/bigquery/docs/reference/v2/jobs}
}
\description{
This is a low-level function that creates a query job. To wait until it
is finished and then retrieve the results, see \code{\link{query_exec}}
}
\seealso{
API documentation for insert method:
  \url{https://developers.google.com/bigquery/docs/reference/v2/jobs/insert}

Other jobs: \code{\link{get_job}};
  \code{\link{insert_upload_job}}; \code{\link{wait_for}}
}

