% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{insert_table}
\alias{insert_table}
\title{Insert empty table}
\usage{
insert_table(project, dataset, table, body = list())
}
\arguments{
\item{project}{project containing this table}

\item{dataset}{dataset containing this table}

\item{table}{name of the table}

\item{body}{A list merged into the request body generated by bigrquery.
This allows you to supply arbitrary additional parameters to the API.}
}
\description{
Insert empty table
}
\seealso{
API documentation:
\url{https://developers.google.com/bigquery/docs/reference/v2/tables/insert}
}
