% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.r
\name{insert_table_job}
\alias{insert_table_job}
\title{Create a new table.}
\usage{
insert_table_job(project, dataset, tablename, schema.fields,
  friendlyName = NULL, description = NULL, timePartitioning.type = NULL,
  timePartitioning.expirationMs = NULL)
}
\arguments{
\item{project}{project name}

\item{dataset}{dataset name}

\item{tablename}{table name}

\item{schema.fields}{Fields description of the table. It needs to be a data.frame with the following columns:
name : (Required) The field name. The name must contain only letters (a-z, A-Z), numbers (0-9), or underscores (_), and must start with a letter or underscore. The maximum length is 128 characters.
type : (Required) The field data type. Possible values include STRING, BYTES, INTEGER, FLOAT, BOOLEAN, TIMESTAMP, DATE, TIME, DATETIME
mode : (Optional) The field mode. Possible values include NULLABLE, REQUIRED and REPEATED. The default value is NULLABLE.
description : (Optional) The field description. The maximum length is 16K characters.}

\item{description}{(Optional) A user-friendly description of this table.}

\item{timePartitioning.type}{(Optional) The only type supported is DAY, which will generate one partition per day based on data loading time.}

\item{timePartitioning.expirationMs}{(Optional) Number of milliseconds for which to keep the storage for a partition.}

\item{frienldyName}{(Optional) A descriptive name for this table.}
}
\description{
This is a low-level function that creates a new table.
}
\seealso{
API documentation for tables method:
  \url{https://cloud.google.com/bigquery/docs/reference/v2/tables}

Other jobs: \code{\link{get_job}},
  \code{\link{insert_query_job}},
  \code{\link{insert_upload_job}}, \code{\link{wait_for}}
}

