% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{is_set_access_cred}
\alias{is_set_access_cred}
\title{Checks if access credentials are set}
\usage{
is_set_access_cred()
}
\value{
TRUE if credentials are set
}
\description{
Checks if access credentials are set
}
