% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-job-api.R
\name{job-API}
\alias{job-API}
\alias{job-API}
\alias{bq_job_meta}
\alias{job-API}
\alias{bq_job_status}
\alias{job-API}
\alias{bq_job_statistics}
\alias{job-API}
\alias{bq_job_wait}
\title{This methods are all wrapper around a single API endpoint: \code{get}.}
\usage{
bq_job_meta(x, fields = NULL)

bq_job_status(x)

bq_job_statistics(x)

bq_job_wait(x, quiet = getOption("bigrquery.quiet"), pause = 0.5)
}
\arguments{
\item{x}{A \link{bq_job}}

\item{fields}{An optional field specification for
\href{https://cloud.google.com/bigquery/docs/api-performance#partial-response}{partial response}}

\item{quiet}{if \code{FALSE} print informative progress messages, if
\code{TRUE} is silent, if \code{NA} displays messages only for long-running
jobs.}

\item{pause}{amount of time to wait between status requests}
}
\description{
This methods are all wrapper around a single API endpoint: \code{get}.
}
\section{API documentation}{

\itemize{
\item \href{https://developers.google.com/bigquery/docs/reference/v2/jobs/get}{meta}
}
}

\examples{
if (bq_testable()) {
jobs <- bq_project_jobs(bq_test_project())
jobs[[1]]

str(bq_job_meta(jobs[[1]]))

# This will wait for a job to complete: if it's already complete
# it will display some status information
bq_job_wait(jobs[[1]])

}
}
