% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr.r
\name{join}
\alias{cross_join}
\alias{cross_join.tbl_bigquery}
\alias{join}
\title{Cross-Joins two tbls together.}
\usage{
cross_join(x, y, copy = FALSE, ...)

\method{cross_join}{tbl_bigquery}(x, y, copy = FALSE, ...)
}
\arguments{
\item{x,y}{tbls to join}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{...}{other parameters passed onto methods}
}
\description{
Cross-Joins two tbls together.
}

