% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.r
\name{load_csv_from_cs}
\alias{load_csv_from_cs}
\title{Load a CSV file from Google Cloud Storage into a table}
\usage{
load_csv_from_cs(project, dataset, table, source_uris, autodetect = TRUE,
  skip_leading_rows = 1, create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_EMPTY", quiet = getOption("bigrquery.quiet"),
  ...)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{table}{The name of the table to load data into}

\item{source_uris}{The fully-qualified URIs that point to a CSV file in
Google Cloud Storage; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.sourceUris}{the API documentation}
for more information}

\item{autodetect}{boolean, default to \code{TRUE} to tell BigQuery to autodetect table schema}

\item{skip_leading_rows}{The number of rows at the top of a CSV file that
BigQuery will skip when loading the data, default to 1}

\item{create_disposition}{Behavior for table creation if the destination
already exists. defaults to \code{"CREATE_IF_NEEDED"},
the only other supported value is \code{"CREATE_NEVER"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.createDisposition}{the API documentation}
for more information}

\item{write_disposition}{Behavior for writing data if the destination already
exists. defaults to \code{"WRITE_EMPTY"}, other possible values are
\code{"WRITE_TRUNCATE"} and \code{"WRITE_APPEND"}; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.writeDisposition}{the API documentation}
for more information}

\item{quiet}{if \code{FALSE}, prints informative status messages.}

\item{...}{Additional arguments merged into the body of the
request. \code{snake_case} will automatically be converted into
\code{camelCase} so you can use consistent argument names.}
}
\description{
Submit an insert job, waits for it, and invisibly returns information on the
destination table
}
\examples{
\dontrun{
project <- "fantastic-voyage-389" # put your project ID here
ds <- "test_dataset" # put your dataset name here
load_csv_from_cs(project = project, dataset = ds, table = "shakespeare",
                 source_uris = "gs://bigrquery_test/shakespeare.csv")
}
}
