% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{load_table_from_cs}
\alias{load_table_from_cs}
\title{Load data from Google Cloud Storage to a table}
\usage{
load_table_from_cs(project, dataset, table, source_uris, ...)
}
\arguments{
\item{project}{project containing this table}

\item{dataset}{dataset containing this table}

\item{table}{name of table to load data into}

\item{source_uris}{the fully-qualified URIs that point to your data in Google
Cloud Storage; see
\href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.sourceUris}{the API documentation}
for more information}
}
\value{
information on the destination table for further consumption by the
  list_tabledata* functions
}
\description{
Load data from Google Cloud Storage to a table
}

