% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_table}
\alias{parse_table}
\title{Parse a BQ-style identifier into project/dataset/table IDs.}
\usage{
parse_table(table, project_id = NULL)
}
\arguments{
\item{table}{table name}

\item{project_id}{(optional) project ID to use if none is provided
in \code{table}}
}
\value{
a list with \code{project_id}, \code{dataset_id}, and
\code{table_id} components (any of which may be \code{NULL}).
}
\description{
Parse a BQ-style identifier into project/dataset/table IDs.
}

