% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-project-api.R, R/old-project.R
\name{project-API}
\alias{project-API}
\alias{bq_project_datasets}
\alias{bq_project_jobs}
\alias{list_datasets}
\title{Project operations}
\usage{
bq_project_datasets(project, page_size = 100, max_pages = 1, warn = TRUE)

bq_project_jobs(project, page_size = 100, max_pages = 1, warn = TRUE)
}
\arguments{
\item{page_size}{Number of items per page.}

\item{max_pages}{Maximum number of pages to retrieve. Use \code{Inf} to retrieve
all pages (this may take a long time!)}

\item{warn}{If \code{TRUE}, warn when there are unretrieved pages.}
}
\value{
:
\itemize{
\item \code{bq_project_datasets()}: a list of \link{bq_dataset}s
\item \code{bq_project_jobs()}: a list of \link{bq_job}s.
}
}
\description{
There is no accompanying \code{bq_project} object because a project is a simple
string.
}
\section{API documentation}{

\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/rest/v2/datasets/list}{datasets}
\item \href{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs/list}{jobs}
}

One day we might also expose the general \href{https://cloud.google.com/resource-manager/reference/rest/v1/projects}{project metadata}.
}

\examples{
if (bq_authable()) {
bq_project_datasets("bigquery-public-data")
bq_project_datasets("githubarchive")
}
}
