% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_big_query}
\alias{read_big_query}
\title{Read a big query from file}
\usage{
read_big_query(query, ...)
}
\arguments{
\item{query}{Path to a file which contains a query.}

\item{...}{Use to substitute values for parameters in the query.}
}
\description{
This function reads a big query from a file, offering basic interpolation.
}
\examples{
\dontrun{
# select.bigquery.sql file
SELECT fullVisitorId FROM ga_sesssions_@YDAY LIMIT 100

# R code
yday <- gsub("-", "", Sys.Date() - 2)
query_exec(read_big_query("select.sql", YDAY = yday), ...)

}
}

