% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{register_credential_fetcher}
\alias{register_credential_fetcher}
\title{Add a new method for fetching OAuth2 credentials.}
\usage{
register_credential_fetcher(f)
}
\description{
This function will add a function to the list of default mechanisms used
when fetching credentials. The single argument should be a function
which takes an oauth endpoint, an oauth app, and a list of scopes, and
returns either NULL or a valid token.
}
\keyword{internal}

