% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{set_service_token}
\alias{set_service_token}
\title{Uses a service token to fetch OAuth2.0 creds for use with BigQuery.}
\usage{
set_service_token(token_file)
}
\arguments{
\item{token_file}{The absolute or relative path to the file containing the service token, a string}
}
\description{
Uses a service token to fetch OAuth2.0 creds for use with BigQuery.
}
\seealso{
Google API documentation:
  \url{https://developers.google.com/identity/protocols/OAuth2ServiceAccount}
}

