% Generated by roxygen2 (4.0.2): do not edit by hand
\name{src_bigquery}
\alias{src_bigquery}
\title{A bigquery data source.}
\usage{
src_bigquery(project, dataset, billing = project)
}
\arguments{
\item{project}{project id or name}

\item{dataset}{dataset name}

\item{billing}{billing project, if different to \code{project}}
}
\description{
Use \code{src_bigquery} to connect to an existing bigquery dataset,
and \code{tbl} to connect to tables within that database.
}
\examples{
\dontrun{
library(dplyr)

# To run this example, replace billing with the id of one of your projects
# with set up with billing
gh <- src_bigquery("githubarchive", "github", billing = "465736758727")
}
}

