% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table-API}
\alias{table-API}
\alias{bq_table_create}
\alias{bq_table_meta}
\alias{bq_table_exists}
\alias{bq_table_delete}
\alias{bq_table_copy}
\alias{bq_table_upload}
\title{Manipulate BigQuery tables}
\usage{
bq_table_create(x, ...)

bq_table_meta(x)

bq_table_exists(x)

bq_table_delete(x)

bq_table_copy(src, dest, create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_EMPTY", billing = NULL, quiet = NA, ...)

bq_table_upload(x, values, billing = NULL,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_APPEND", quiet = NA, ...)
}
\arguments{
\item{x}{A \link{bq_table}, or an object coercible to a \code{bq_table}.}

\item{src, dest}{Source and desintation \link{bq_table}s.}

\item{create_disposition}{Behavior if the destination table does not already
exist. Default will create table on demand; set to \code{"CREATE_NEVER"} to
never create (and only copy into existing table).}

\item{write_disposition}{Behavior if the destination already exists.
The default, \code{"WRITE_EMPTY"}, will error if the table already contains
data. Other possible values are \code{"WRITE_TRUNCATE"} and \code{"WRITE_APPEND"}.}

\item{billing}{Project to bill. Defaults to the \code{dest} project.}
}
\value{
\itemize{
\item \code{bq_table_get()}: a \href{https://developers.google.com/bigquery/docs/reference/v2/tables}{table resource list}
\item \code{bq_table_get()}: either \code{TRUE} or \code{FALSE}
\item \code{bq_table_copy()}: a table reference to the new table
}
}
\description{
Manipulate BigQuery tables
}
\section{API documentation}{

\itemize{
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/insert}{create}
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/get}{get}
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/delete}{delete}
\item \href{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs#configuration.copy}{copy}
}
}

\examples{
if (bq_testable()) {
ds <- bq_test_dataset()

bq_mtcars <- bq_table(ds, "mtcars")
bq_table_exists(bq_mtcars)
str(bq_table_meta(bq_mtcars))

bq_table_upload(bq_mtcars, mtcars)
bq_table_exists(bq_mtcars)

bq_table_delete(bq_mtcars)
bq_table_exists(bq_mtcars)

my_natality <- bq_table(ds, "mynatality")
bq_table_copy("publicdata.samples.natality", my_natality)
}
}
