% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{update_dataset}
\alias{update_dataset}
\title{Updates an existing dataset in a project}
\usage{
update_dataset(project, dataset, description = NULL, friendlyName = NULL)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The dataset to update, a string}

\item{description}{The dataset description, a string}

\item{friendlyName}{The dataset's friendly name, a string}
}
\description{
Updates an existing dataset in a project
}
\examples{
\dontrun{
update_dataset("myproject", "existing_dataset", "my description", "friendly name")
}
}
\seealso{
Google API documentation:
  \url{https://cloud.google.com/bigquery/docs/reference/v2/datasets/update}

Other datasets: \code{\link{delete_dataset}},
  \code{\link{get_dataset}}, \code{\link{insert_dataset}},
  \code{\link{list_datasets}}
}

