% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{bigrquery}
\alias{bigquery}
\alias{bigrquery}
\alias{bigrquery-package}
\title{An R interface to google's bigquery.}
\description{
For more information about how bigrquery works, and how to get started,
please see the project development page: \url{github.com/hadley/bigrquery}.
The most important method to get started with is \code{\link{query_exec}}.
}

