% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-stream.R
\name{bq_table_stream}
\alias{bq_table_stream}
\title{Streams data from data.from into BigQUery}
\usage{
bq_table_stream(x, values, skip_invalid_rows = "false",
  ignore_unknown_values = "false", template_suffix = NULL)
}
\arguments{
\item{x}{bq_table where values will be streamed to}

\item{values}{data.frame that will be streamed}

\item{skip_invalid_rows}{(optional) Insert all valid rows of a request,
even if invalid rows exist. The default value is False, which
causes the entire request to fail if any invalid rows exist.}

\item{ignore_unknown_values}{(optional) Accept rows that contain values that do not match the
schema. The unknown values are ignored. Default is False, which
treats unknown values as errors.}

\item{template_suffix}{(optional) treat \code{name} as a template table and provide a suffix.
BigQuery will create the table \code{<name> + <template_suffix>} based
on the schema of the template table. See
https://cloud.google.com/bigquery/streaming-data-into-bigquery#template-tables}
}
\description{
Streams data from data.from into BigQUery
}
\seealso{
https://cloud.google.com/bigquery/streaming-data-into-bigquery
}
