% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-stream.R
\name{bq_table_stream}
\alias{bq_table_stream}
\title{Stream data.frame into BigQUery}
\usage{
bq_table_stream(
  x,
  values,
  insert_ids = NULL,
  skip_invalid_rows = FALSE,
  ignore_unknown_values = FALSE,
  template_suffix = NULL
)
}
\arguments{
\item{x}{bq_table where values will be streamed to}

\item{values}{data.frame that will be streamed}

\item{insert_ids}{(optional) Unique ids, one per row being inserted.
If omitted, unique IDs are created.}

\item{skip_invalid_rows}{(optional) Insert all valid rows of a request,
even if invalid rows exist. The default value is FALSE, which
causes the entire request to fail if any invalid rows exist.}

\item{ignore_unknown_values}{(optional) Accept rows that contain values that do not match the
schema. The unknown values are ignored. Default is FALSE, which
treats unknown values as errors.}

\item{template_suffix}{(optional) treat \code{name} as a template table and provide a suffix.
BigQuery will create the table \verb{<name> + <template_suffix>} based
on the schema of the template table. See
https://cloud.google.com/bigquery/streaming-data-into-bigquery#template-tables}
}
\value{
NULL if there are no errors, otherwise list of errors with reasons
}
\description{
Stream data.frame into BigQUery
}
\examples{
\dontrun{
ds <- bq_test_dataset()
tb <- bq_table(ds, "cars")
# create an empty table with schema for `cars` data.frame
bq_table_create(tb, cars)

bq_table_stream(tb, cars))
}
}
\seealso{
https://cloud.google.com/bigquery/streaming-data-into-bigquery
}
