% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-auth.R
\name{bq_auth_config}
\alias{bq_auth_config}
\alias{bq_oauth_app}
\title{View or edit auth config}
\usage{
bq_auth_config(app = NULL, path = NULL)

bq_oauth_app()
}
\arguments{
\item{app}{OAuth app. Defaults to a tidyverse app.}

\item{path}{Path to the JSON file.}
}
\value{
\code{bq_auth_config()}: An object of R6 class \code{AuthState},
which is defined in the gargle package. \code{bq_oauth_app()}: the
current \code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
View or edit auth config

\code{bq_oauth_app()} retrieves the currently configured OAuth app.
}
\examples{
## this will print current config
bq_auth_config()

if (require(httr)) {
  ## bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  bq_auth_config(app = google_app)
}

\dontrun{
## bring your own app via JSON downloaded from Google Developers Console
bq_auth_config(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}
}
\seealso{
Other auth functions: \code{\link{bq_auth}}
}
\concept{auth functions}
