% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-auth.R
\name{bq_deauth}
\alias{bq_deauth}
\title{Clear current token}
\usage{
bq_deauth()
}
\description{
Clears any currently stored token. The next time bigrquery needs a token, the
token acquisition process starts over, with a fresh call to \code{\link[=bq_auth]{bq_auth()}} and,
therefore, internally, a call to \code{\link[gargle:token_fetch]{gargle::token_fetch()}}. Unlike some other
packages that use gargle, bigrquery is not usable in a de-authorized state.
Therefore, calling \code{bq_deauth()} only clears the token, i.e. it does NOT
imply that subsequent requests are made with an API key in lieu of a token.
}
\examples{
\dontrun{
bq_deauth()
}
}
\seealso{
Other auth functions: 
\code{\link{bq_auth_configure}()},
\code{\link{bq_auth}()}
}
\concept{auth functions}
