% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-field.R
\name{bq_field}
\alias{bq_field}
\alias{bq_fields}
\title{BiqQuery field (and fields) class}
\usage{
bq_field(name, type, mode = "nullable", fields = NULL)

bq_fields(x)
}
\arguments{
\item{name}{Field name}

\item{type}{Field type}

\item{mode}{Field mode}

\item{fields}{For a field of type "record", a list of sub-fields.}

\item{x}{A list of \code{bg_fields}}
}
\description{
BiqQuery field (and fields) class
}
