% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-auth.R
\name{bq_has_token}
\alias{bq_has_token}
\title{Is there a token on hand?}
\usage{
bq_has_token()
}
\value{
Logical.
}
\description{
Reports whether bigrquery has stored a token, ready for use in downstream
requests.
}
\examples{
bq_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{bq_token}()}
}
\concept{low-level API functions}
