% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep-projects.R, R/projects.r
\name{list_projects}
\alias{list_projects}
\alias{bq_projects}
\title{List all projects to which you have been granted any project role.}
\usage{
list_projects()

bq_projects()
}
\value{
A character vector of project ids named with their friendly names.
}
\description{
You can also work with \href{https://cloud.google.com/bigquery/public-data/}{public datasets}
but you will need to provide a \code{billing} project whenever you perform
any non-free operation.
}
\examples{
if (bq_authable()) {
bq_projects()
}
}
\seealso{
\href{https://cloud.google.com/bigquery/docs/reference/v2/projects/list}{API documentation}
}
