% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-projects.r, R/old-projects.R
\name{bq_projects}
\alias{bq_projects}
\alias{list_projects}
\title{Projects API}
\usage{
bq_projects()

list_projects()
}
\value{
A character vector of project ids named with their friendly name
(if set).
}
\description{
There is currently only one project method which returns all projects to
which you have been granted access. You can also work with
\href{https://cloud.google.com/bigquery/public-data/}{public datasets},
but you will need to provide a \code{billing} project whenever you perform
any non-free operation.
}
\details{
There is no accompanying \code{bq_project} object because a project is a simple
string.
}
\examples{
if (bq_authable()) {
bq_projects()
}
}
\seealso{
\href{https://cloud.google.com/bigquery/docs/reference/v2/projects/list}{API documentation}
}
