% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{bq_table}
\alias{bq_table}
\alias{as_bq_table}
\title{Referencd to a BigQuery table}
\usage{
bq_table(project, dataset, table)

as_bq_table(x)
}
\description{
\code{bq_table()} creates from individual \code{project}, \code{dataset}, and \code{table}
components; \code{as_bq_table()} will create from a character vector or a list.
See \link{table-API} for operations that work with the remote table.
}
\examples{
natality <- bq_table("publicdata", "samples", "natality")
natality

as_bq_table("publicdata.samples.natality")

# Included primarily for backward compatbility
as_bq_table(list(
  project_id = "publicdata",
  dataset_id = "samples",
  table_id = "natality"
))
}
