% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-download.R
\name{bq_table_download}
\alias{bq_table_download}
\title{Download table data}
\usage{
bq_table_download(x, max_results = Inf, page_size = 10000,
  start_index = 0L)
}
\arguments{
\item{x}{A \link{bq_table}}

\item{max_results}{Maximum nubmer of results to return}

\item{page_size}{The number of rows returned per page. Make this smaller
if you have many fields or large records and you are seeing a
"responseTooLarge" error.}

\item{max_pages}{Maximum number of pages to retrieve. Use \code{Inf} retrieve all
rows.}
}
\description{
Download table data
}
\section{API documentation}{

\itemize{
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tabledata/list}{list}
}
}

\examples{
if (bq_testable()) {
df <- bq_table_download("publicdata.samples.natality", max_results = 35000)
}
}
