% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-test.R
\name{bq_test_project}
\alias{bq_test_project}
\title{Project to use for testing bigquery}
\usage{
bq_test_project()
}
\description{
You'll need to set up the \code{BIGQUERY_TEST_PROJECT} env var if you want to
run bigquery tests locally. I recommend creating a new project because
the tests involve both reading and writing. You will also need to
have billing billing enabled for the project.
}
\keyword{internal}
