% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{delete_table}
\alias{delete_table}
\title{Delete a table.}
\usage{
delete_table(project, dataset, table)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{table}{name of the table}
}
\description{
Delete a table.
}
\examples{
\dontrun{
get_table("publicdata", "samples", "natality")
}
}
\seealso{
API documentation:
\url{https://developers.google.com/bigquery/docs/reference/v2/tables/delete}

Other tables: \code{\link{get_table}},
  \code{\link{list_tables}}
}
