% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{format_dataset}
\alias{format_dataset}
\title{Format dataset and project ID as a BQ-style identifier}
\usage{
format_dataset(project_id, dataset)
}
\arguments{
\item{project_id}{project ID}

\item{dataset}{dataset name}
}
\value{
a character.
}
\description{
This function composes a dataset identifier from its individual
components.
}
\seealso{
Other identifier functions: \code{\link{format_table}},
  \code{\link{parse_dataset}}, \code{\link{parse_table}}
}
