% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.r
\name{get_job}
\alias{get_job}
\title{Check status of a job.}
\usage{
get_job(project, job)
}
\arguments{
\item{project}{project name}

\item{job}{job id}
}
\value{
a job resource list, as documented at
\url{https://developers.google.com/bigquery/docs/reference/v2/jobs}
}
\description{
Check status of a job.
}
\seealso{
API documentation for get method:
\url{https://developers.google.com/bigquery/docs/reference/v2/jobs/get}

\code{\link[=wait_for]{wait_for()}} to wait for a job to complete

Other jobs: \code{\link{insert_extract_job}},
  \code{\link{insert_load_csv_from_cs_job}},
  \code{\link{insert_query_job}},
  \code{\link{insert_upload_job}}, \code{\link{wait_for}}
}
