% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-job.R
\name{get_job}
\alias{get_job}
\title{Check status of a job \link{deprecated}}
\usage{
get_job(project, job)
}
\arguments{
\item{project}{project name}

\item{job}{job id}
}
\value{
a job resource list, as documented at
\url{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs}
}
\description{
Please use \link{api-job} instead
}
\seealso{
API documentation for get method:
\url{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs/get}

\code{\link[=wait_for]{wait_for()}} to wait for a job to complete

Other jobs: 
\code{\link{insert_extract_job}()},
\code{\link{insert_query_job}()},
\code{\link{insert_upload_job}()},
\code{\link{wait_for}()}
}
\concept{jobs}
\keyword{internal}
