% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.r
\name{get_table}
\alias{get_table}
\title{Retrieve table metadata}
\usage{
get_table(project, dataset, table)
}
\arguments{
\item{project}{project containing this table}

\item{dataset}{dataset containing this table}

\item{table}{name of the table}
}
\value{
A table resource list, as described by
 \url{https://developers.google.com/bigquery/docs/reference/v2/tables}
}
\description{
Retrieve table metadata
}
\examples{
\dontrun{
get_table("publicdata", "samples", "natality")
get_table("githubarchive", "github", "timeline")
}
}
\seealso{
API documentation:
 \url{https://developers.google.com/bigquery/docs/reference/v2/tables/get}

Other tables: \code{\link{delete_table}};
  \code{\link{list_tables}}
}

