% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{insert_dataset}
\alias{insert_dataset}
\title{Creates a new dataset in a project}
\usage{
insert_dataset(project, dataset, description = NULL, friendlyName = NULL)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{description}{The dataset description, a string}

\item{friendlyName}{The dataset's friendly name, a string}
}
\description{
Creates a new dataset in a project
}
\examples{
\dontrun{
insert_dataset("myproject", "new_dataset")
}
}
\seealso{
Google API documentation:
  \url{https://cloud.google.com/bigquery/docs/reference/v2/datasets/insert}

Other datasets: \code{\link{delete_dataset}},
  \code{\link{get_dataset}}, \code{\link{list_datasets}},
  \code{\link{update_dataset}}
}

