% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{insert_table}
\alias{insert_table}
\title{Insert empty table}
\usage{
insert_table(project, dataset, table, ...)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{table}{name of the table}

\item{...}{Additional arguments merged into the body of the
request. \code{snake_case} will automatically be converted into
\code{camelCase} so you can use consistent argument names.}
}
\description{
Insert empty table
}
\seealso{
API documentation:
\url{https://developers.google.com/bigquery/docs/reference/v2/tables/insert}
}
