% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_upload_job}
\alias{insert_upload_job}
\title{Upload data.}
\usage{
insert_upload_job(project, dataset, table, values, billing = project)
}
\arguments{
\item{project}{project name}

\item{dataset}{dataset name}

\item{table}{name of table to insert values into}

\item{billing}{project to bill to, if different to \code{project}}

\item{value}{data frame of data to upload}
}
\description{
This sends all of the data inline in the HTTP request so is only suitable
for relatively small datasets.
}
\examples{
\donttest{
list_datasets("193487687779")
list_tables("193487687779", "houston")
job <- insert_upload_job("193487687779", "houston", "mtcars", mtcars)
wait_for(job)
list_tables("193487687779", "houston")
delete_table("193487687779", "houston", "mtcars")
}
}
\seealso{
Google API documentation:
\url{https://developers.google.com/bigquery/loading-data-into-bigquery#loaddatapostrequest}

Other jobs: \code{\link{get_job}};
  \code{\link{insert_query_job}}; \code{\link{wait_for}}
}

