% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.r
\name{list_datasets}
\alias{list_datasets}
\title{List the datasets in a project}
\usage{
list_datasets(project)
}
\arguments{
\item{project}{The project name, a string}
}
\value{
a character vector of dataset names
}
\description{
List the datasets in a project
}
\examples{
\dontrun{
list_datasets("publicdata")
list_datasets("githubarchive")
}
}
\seealso{
Google API documentation:
  \url{https://developers.google.com/bigquery/docs/reference/v2/datasets/list}
}

