% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{list_datasets}
\alias{list_datasets}
\title{List the datasets in a project}
\usage{
list_datasets(project, page_size = 50, max_pages = Inf)
}
\arguments{
\item{project}{The project name, a string}

\item{page_size}{Number of items per page}

\item{max_pages}{Maximum number of pages to retrieve}
}
\value{
a character vector of dataset names
}
\description{
List the datasets in a project
}
\examples{
\dontrun{
list_datasets("publicdata")
list_datasets("githubarchive")
}
}
\seealso{
Google API documentation:
\url{https://developers.google.com/bigquery/docs/reference/v2/datasets/list}

Other datasets: \code{\link{delete_dataset}},
  \code{\link{get_dataset}}, \code{\link{insert_dataset}},
  \code{\link{update_dataset}}
}
