% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-project.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets \link{deprecated}}
\usage{
list_datasets(project, page_size = 50, max_pages = Inf)
}
\description{
Please use \link{api-project} instead.
}
\keyword{internal}
