% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.r
\name{list_projects}
\alias{list_projects}
\title{List all projects to which you have been granted any project role.}
\usage{
list_projects()
}
\value{
a character vector of project ids named with their friendly names.
}
\description{
List all projects to which you have been granted any project role.
}
\examples{
\dontrun{
list_projects()
}
}
\seealso{
API documentation at
\url{https://developers.google.com/bigquery/docs/reference/v2/projects/list}
}
