% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projects.r
\name{list_projects}
\alias{list_projects}
\title{List projects that you have access to}
\usage{
list_projects()
}
\value{
a character vector of project ids named with their friendly names
  Always includes publicly available sampledata project.
}
\description{
List projects that you have access to
}
\examples{
\dontrun{
list_projects()
}
}
\seealso{
API documentation at
  \url{https://developers.google.com/bigquery/docs/reference/v2/projects/list}
}

