% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{list_tables}
\alias{list_tables}
\title{List available tables in dataset.}
\usage{
list_tables(project, dataset, page_size = 50, max_pages = Inf)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{page_size}{Number of items per page}

\item{max_pages}{Maximum number of pages to retrieve}
}
\value{
a character vector of table names
}
\description{
List available tables in dataset.
}
\examples{
\dontrun{
list_tables("publicdata", "samples")
list_tables("githubarchive", "github")
list_tables("publicdata", "samples", max_pages = 2, page_size = 2)
}
}
\seealso{
API documentation:
\url{https://developers.google.com/bigquery/docs/reference/v2/tables/list}

Other tables: \code{\link{delete_table}},
  \code{\link{get_table}}
}
