% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_dataset}
\alias{parse_dataset}
\title{Parse a BQ-style identifier into project/dataset IDs.}
\usage{
parse_dataset(dataset, project_id = NULL)
}
\arguments{
\item{dataset}{dataset name}

\item{project_id}{(optional) project ID to use if none is provided
in \code{dataset}}
}
\value{
a list with \code{project_id} and \code{dataset_id} components
(either of which may be \code{NULL}).
}
\description{
Parse a BQ-style identifier into project/dataset IDs.
}

