% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{parse_table}
\alias{parse_table}
\title{Parse a BQ-style identifier into project/dataset/table IDs}
\usage{
parse_table(table, project_id = NULL)
}
\arguments{
\item{table}{table name}

\item{project_id}{(optional) project ID to use if none is provided
in \code{table}}
}
\value{
a list with \code{project_id}, \code{dataset_id}, and
\code{table_id} components (any of which may be \code{NULL}).
}
\description{
This function splits a table identifier (given as character) into its
components.
}
\seealso{
Other identifier functions: \code{\link{format_dataset}},
  \code{\link{format_table}}, \code{\link{parse_dataset}}
}

