% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-project-api.R, R/old-project.R
\name{project-API}
\alias{project-API}
\alias{bq_project_datasets}
\alias{bq_project_jobs}
\alias{list_datasets}
\title{Project operations}
\usage{
bq_project_datasets(project, page_size = 50, max_pages = 1)

bq_project_jobs(project, page_size = 50, max_pages = 1)
}
\arguments{
\item{page_size, max_pages}{Pagination parameters.}
}
\value{
:
\itemize{
\item \code{bq_project_dataset()}: a list of \link{bq_dataset}s
\item \code{bq_project_jobs()}: a list of \link{bq_job}s.
\item \code{bq_project_query()}: a \link{bq_table}.
}
}
\description{
Project operations
}
\section{API documentation}{

\itemize{
\item \href{https://cloud.google.com/bigquery/docs/reference/rest/v2/datasets/list}{datasets}
\item \href{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs/list}{jobs}
}

One day we might also expose the general \href{https://cloud.google.com/resource-manager/reference/rest/v1/projects}{project metadata}.
}

\examples{
if (bq_authable()) {
bq_project_datasets("bigquery-public-data")
bq_project_datasets("githubarchive")
}
}
