% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{set_service_token}
\alias{set_service_token}
\title{Sets the service token to use when fetching the OAuth token
creds. This will throw an error if the file does not exist,
we do not have permissions to read it, or if it is not valid JSON.}
\usage{
set_service_token(token_file)
}
\arguments{
\item{The}{absolute or relative path to the file containing the service token, a string}
}
\description{
Sets the service token to use when fetching the OAuth token
creds. This will throw an error if the file does not exist,
we do not have permissions to read it, or if it is not valid JSON.
}
\seealso{
Google API documentation:
  \url{https://developers.google.com/identity/protocols/OAuth2ServiceAccount}
}

