% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-table-api.R
\name{table-API}
\alias{table-API}
\alias{bq_table_create}
\alias{bq_table_meta}
\alias{bq_table_size}
\alias{bq_table_exists}
\alias{bq_table_delete}
\alias{bq_table_copy}
\alias{bq_table_upload}
\alias{bq_table_fields}
\alias{bq_table_download}
\alias{bq_table_extract}
\title{Manipulate BigQuery tables}
\usage{
bq_table_create(x, ...)

bq_table_meta(x, fields = NULL)

bq_table_size(x)

bq_table_exists(x)

bq_table_delete(x)

bq_table_copy(src, dest, ..., quiet = NA)

bq_table_upload(x, values, ..., quiet = NA)

bq_table_fields(x, fields = NULL)

bq_table_download(x, billing = NULL, ..., page_size = 10000,
  max_pages = 10, warn = TRUE, quiet = NA)

bq_table_extract(x, urls, ..., quiet = NA)
}
\arguments{
\item{x}{A \link{bq_table}, or an object coercible to a \code{bq_table}.}

\item{src, dest}{Source and desintation \link{bq_table}s.}

\item{quiet}{if \code{FALSE} print informative progress messages, if
\code{TRUE} is silent, if \code{NA} displays messages for long-running
jobs.}

\item{billing}{Project to bill. Defaults to the \code{dest} project.}

\item{create_disposition}{Behavior if the destination table does not already
exist. Default will create table on demand; set to \code{"CREATE_NEVER"} to
never create (and only copy into existing table).}

\item{write_disposition}{Behavior if the destination already exists.
The default, \code{"WRITE_EMPTY"}, will error if the table already contains
data. Other possible values are \code{"WRITE_TRUNCATE"} and \code{"WRITE_APPEND"}.}
}
\value{
\itemize{
\item \code{bq_table_get()}: a \href{https://developers.google.com/bigquery/docs/reference/v2/tables}{table resource list}
\item \code{bq_table_get()}: either \code{TRUE} or \code{FALSE}
\item \code{bq_table_copy()}: a table reference to the new table
}
}
\description{
Manipulate BigQuery tables
}
\section{API documentation}{

\itemize{
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/insert}{insert}
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/get}{get}
\item \href{https://developers.google.com/bigquery/docs/reference/v2/tables/delete}{delete}
}
}

\examples{
if (bq_testable()) {
ds <- bq_test_dataset()

bq_mtcars <- bq_table(ds, "mtcars")
bq_table_exists(bq_mtcars)
str(bq_table_meta(bq_mtcars))

bq_table_upload(bq_mtcars, mtcars)
bq_table_exists(bq_mtcars)

bq_table_delete(bq_mtcars)
bq_table_exists(bq_mtcars)

my_natality <- bq_table(ds, "mynatality")
bq_table_copy("publicdata.samples.natality", my_natality)
}
}
