% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{table-API}
\alias{table-API}
\alias{bq_table_create}
\alias{bq_table_get}
\alias{bq_table_exists}
\alias{bq_table_delete}
\alias{bq_table_copy}
\alias{bq_table_upload}
\title{Manipulate tables}
\usage{
bq_table_create(x, ...)

bq_table_get(x)

bq_table_exists(x)

bq_table_delete(x)

bq_table_copy(src, dest, create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_EMPTY", billing = NULL, quiet = NA, ...)

bq_table_upload(x, values, billing = NULL,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_APPEND", quiet = NA, ...)
}
\arguments{
\item{x}{A \link{bq_table}}

\item{src, dest}{Source and desintation \link{bq_table}s.}

\item{create_disposition}{Behavior if the destination table does not already
exist. Default will create table on demand; set to \code{"CREATE_NEVER"} to
never create (and only copy into existing table).}

\item{write_disposition}{Behavior if the destination already exists.
The default, \code{"WRITE_EMPTY"}, will error if the table already contains
data. Other possible values are \code{"WRITE_TRUNCATE"} and \code{"WRITE_APPEND"}.}

\item{billing}{Project to bill. Defaults to the \code{dest} project.}
}
\value{
\itemize{
\item \code{bq_table_get()}: a \href{https://developers.google.com/bigquery/docs/reference/v2/tables}{table resource list}
\item \code{bq_table_get()}: either \code{TRUE} or \code{FALSE}
\item \code{bq_table_copy()}: a table reference to the new table
}
}
\description{
Manipulate tables
}
\examples{
\dontrun{
insert_dataset(bq_test_project(), "table_api")

bq_mtcars <- bq_table(bq_test_project(), "table_api", "mtcars")
bq_table_exists(bq_mtcars)

bq_table_upload(bq_mtcars, mtcars)
bq_table_exists(bq_mtcars)

bq_table_delete(bq_mtcars)
bq_table_exists(bq_mtcars)

my_natality <- bq_table(bq_test_project(), "table_api", "mynatality")
bq_table_copy("publicdata:samples:natality", my_natality)

delete_dataset(bq_test_project(), "table_api", deleteContents = TRUE)
}
}
