% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-table.R
\name{table-dep}
\alias{table-dep}
\alias{insert_table}
\alias{get_table}
\alias{exists_table}
\alias{delete_table}
\title{Table API \link{deprecated}}
\usage{
insert_table(project, dataset, table, ...)

get_table(project, dataset, table)

exists_table(project, dataset, table)

delete_table(project, dataset, table)
}
\description{
See \link{api-table} for recommended API.
}
\keyword{internal}
