% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{update_dataset}
\alias{update_dataset}
\title{Updates an existing dataset in a project}
\usage{
update_dataset(project, dataset, ...)
}
\arguments{
\item{project}{The project name, a string}

\item{dataset}{The name of the dataset to create, a string}

\item{...}{Additional arguments merged into the body of the
request. \code{snake_case} will automatically be converted into
\code{camelCase} so you can use consistent argument names.}
}
\description{
Updates an existing dataset in a project
}
\examples{
\dontrun{
update_dataset("myproject", "existing_dataset", "my description", "friendly name")
}
}
\seealso{
Google API documentation:
\url{https://cloud.google.com/bigquery/docs/reference/v2/datasets/update}

Other datasets: \code{\link{delete_dataset}},
  \code{\link{get_dataset}}, \code{\link{insert_dataset}},
  \code{\link{list_datasets}}
}
