% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-job.R
\name{wait_for}
\alias{wait_for}
\title{Wait for a job to complete \link{deprecated}}
\usage{
wait_for(job, quiet = getOption("bigrquery.quiet"), pause = 0.5)
}
\arguments{
\item{job}{job to wait for. Probably result of \code{\link[=insert_query_job]{insert_query_job()}}
or \code{\link[=insert_upload_job]{insert_upload_job()}}}

\item{quiet}{if \code{FALSE} print informative progress messages, if
\code{TRUE} is silent, if \code{NA} displays messages for long-running
jobs.}

\item{pause}{amount of time to wait between status requests}
}
\description{
Please use \link{api-job} instead
}
\seealso{
Other jobs: 
\code{\link{get_job}()},
\code{\link{insert_extract_job}()},
\code{\link{insert_query_job}()},
\code{\link{insert_upload_job}()}
}
\concept{jobs}
\keyword{internal}
